+++ Estimation/BamProcessing/BamProcessor.cpp
@@ -12,8 +12,8 @@ namespace Estimation
 		BamProcessor::BamProcessor(CellsDataContainer &container, const BamTags &tags, bool print_bam)
 			: BamProcessorAbstract(tags)
 			, _container(container)
-			, print_bam(print_bam)
-			, total_intergenic_reads(0)
+			, _print_bam(print_bam)
+			, _total_intergenic_reads(0)
 		{}
 
 		void BamProcessor::save_read(const std::string& cell_barcode, const std::string& chr_name, const std::string& umi,
@@ -21,7 +21,7 @@ namespace Estimation
 		{
 			if (gene == "")
 			{
-				this->total_intergenic_reads++;
+				this->_total_intergenic_reads++;
 			}
 
 			this->_container.add_record(cell_barcode, umi, gene, chr_name, umi_mark);
@@ -35,8 +35,13 @@ namespace Estimation
 				cant_parse_msg << "can't parse " << (100.0*this->cant_parse_reads_num() / this->total_reads_num()) <<"% reads; ";
 			}
 
+			if (this->low_quality_reads_num() > 0)
+			{
+				cant_parse_msg << "low-quality reads: " << (100.0*this->low_quality_reads_num() / this->total_reads_num()) <<"%; ";
+			}
+
 			L_TRACE << trace_prefix << ": " << this->total_reads_num() << " total reads; " << std::setprecision(3)
-					<< (100.0*this->total_intergenic_reads / this->total_reads_num()) <<"% intergenic; "
+					<< (100.0*this->_total_intergenic_reads / this->total_reads_num()) <<"% intergenic; "
 					<< (100.0*this->container().has_exon_reads_num() / this->total_reads_num()) <<"% touch exon; "
 					<< (100.0*this->container().has_intron_reads_num() / this->total_reads_num()) <<"% touch intron; "
 					<< (100.0*this->container().has_not_annotated_reads_num() / this->total_reads_num()) <<"% touch not annotated regions; "
@@ -46,7 +51,7 @@ namespace Estimation
 
 		void BamProcessor::update_bam(const std::string &bam_file, const BamTools::BamReader &reader)
 		{
-			if (!this->print_bam)
+			if (!this->_print_bam)
 				return;
 
 			BamProcessorAbstract::update_bam(bam_file, reader);
@@ -60,11 +65,10 @@ namespace Estimation
 		void BamProcessor::write_alignment(BamTools::BamAlignment alignment, const std::string &gene,
 										const Tools::ReadParameters &read_params)
 		{
-			if (!this->print_bam)
+			if (!this->_print_bam)
 				return;
 
-			this->save_alignment(alignment, read_params.read_name_safe(), gene,
-								 read_params.cell_barcode(), read_params.umi());
+			this->save_alignment(alignment, read_params, gene);
 		}
 
 		const CellsDataContainer &BamProcessor::container() const