+++ Estimation/BamProcessing/BamProcessorAbstract.cpp
@@ -1,6 +1,8 @@
-#include <Tools/Logs.h>
 #include "BamProcessorAbstract.h"
+
 #include "BamController.h"
+#include <Tools/Logs.h>
+#include <Tools/ReadParameters.h>
 
 namespace Estimation
 {
@@ -9,6 +11,7 @@ namespace BamProcessing
 	BamProcessorAbstract::BamProcessorAbstract(const BamTags &tags_info)
 		: _total_reads_num(0)
 		, _cant_parse_reads_num(0)
+		, _low_quality_reads_num(0)
 		, _tags(tags_info)
 	{}
 
@@ -48,23 +51,33 @@ namespace BamProcessing
 		return this->_cant_parse_reads_num;
 	}
 
-	void BamProcessorAbstract::save_alignment(BamTools::BamAlignment alignment, const std::string &name,
-											  const std::string &gene, const std::string &barcode, const std::string &umi)
+	size_t BamProcessorAbstract::low_quality_reads_num() const
 	{
-		alignment.Name = name;
+		return this->_low_quality_reads_num;
+	}
+
+	void BamProcessorAbstract::save_alignment(BamTools::BamAlignment alignment, const Tools::ReadParameters &params,
+	                                          const std::string &gene)
+	{
 		if (gene != "")
 		{
 			alignment.AddTag(this->_tags.gene, "Z", gene);
 		}
 
-		alignment.AddTag(this->_tags.cb, "Z", barcode);
-		alignment.AddTag(this->_tags.umi, "Z", umi);
+		// TODO: add quality tags
+		alignment.AddTag(this->_tags.cell_barcode, "Z", params.cell_barcode());
+		alignment.AddTag(this->_tags.umi, "Z", params.umi());
 		this->_writer.SaveAlignment(alignment);
 	}
 
 	void BamProcessorAbstract::inc_cant_parse_num()
 	{
 		++this->_cant_parse_reads_num;
+	}
+
+	void BamProcessorAbstract::inc_low_quality_num()
+	{
+		++this->_low_quality_reads_num;
 	}
 }
 }