+++ Estimation/BamProcessing/ReadMapParamsParser.h
@@ -14,12 +14,11 @@ namespace Estimation
 			Tools::reads_params_map_t _reads_params;
 
 		private:
-			void init(const std::string &read_param_filenames, bool save_read_name);
+			void init(const std::string &read_param_filenames);
 
 		public:
-			ReadMapParamsParser(const std::string &gtf_path, bool save_read_names,
-			                    const std::string &read_param_filenames, const BamTags &tags,
-			                    bool gene_in_chromosome_name);
+			ReadMapParamsParser(const std::string &gtf_path, const std::string &read_param_filenames,
+			                    const BamTags &tags, bool gene_in_chromosome_name);
 			virtual ~ReadMapParamsParser();
 
 			virtual bool get_read_params(const BamTools::BamAlignment &alignment, Tools::ReadParameters &read_params) override;