+++ Estimation/BamProcessing/ReadParamsParser.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <unordered_set>
 #include <api/BamAlignment.h>
 
 #include <Estimation/CellsDataContainer.h>
@@ -26,6 +27,14 @@ namespace Estimation
 		protected:
 			const BamTags tags;
 
+		private:
+			UMI::Mark parse_read_type(const BamTools::BamAlignment &alignment, std::string &gene) const;
+			UMI::Mark get_gene_from_reference(const std::string &chr_name, const BamTools::BamAlignment &alignment,
+			                                  std::string &gene) const;
+
+			bool find_exon(Tools::RefGenesContainer::query_results_t query_results,
+			               Tools::RefGenesContainer::QueryResult &exon_result) const;
+
 		public:
 			ReadParamsParser(const std::string &genes_filename, const BamTags &tags, bool gene_in_chromosome_name);
 
@@ -33,13 +42,7 @@ namespace Estimation
 			UMI::Mark get_gene(const std::string &chr_name, BamTools::BamAlignment alignment,
 			                                  std::string &gene) const;
 
-			bool find_exon(Tools::RefGenesContainer::query_results_t query_results,
-			               Tools::RefGenesContainer::QueryResult &exon_result) const;
-
 			bool has_introns() const;
-
-			UMI::Mark get_gene_from_reference(const std::string &chr_name, const BamTools::BamAlignment &alignment,
-			                                  std::string &gene) const;
 		};
 	}
 }