+++ TagsSearch/IndropV3TagsFinder.h
@@ -29,8 +29,8 @@ namespace TagsSearch
 		IndropV3TagsFinder(const std::vector<std::string> &fastq_filenames,
 		                   const boost::property_tree::ptree &barcodes_config,
 		                   const boost::property_tree::ptree &processing_config,
-		                   const std::shared_ptr<TextWriter> &writer,
-		                   bool save_stats);
+		                   const std::shared_ptr<ConcurrentGzWriter> &writer,
+		                   bool save_stats, bool save_read_params);
 
 		std::string parse_umi(const std::string &cb2_seq) const;
 	};