+++ TagsSearch/TagsFinderBase.h
@@ -6,7 +6,7 @@
 #include "Tools/UtilFunctions.h"
 #include <Tools/ScSpConcurrentQueue.h>
 #include <Tools/BlockingConcurrentQueue.h>
-#include "TextWriter.h"
+#include "ConcurrentGzWriter.h"
 
 #include <string>
 
@@ -37,19 +37,19 @@ namespace TagsSearch
 
 	private:
 		const bool _save_stats;
+		const bool _save_read_params;
+		const int _quality_threshold;
 		const std::string _file_uid;
 
 		std::atomic<long> _total_reads_read;
+		std::atomic<long> _low_quality_reads;
 		std::atomic<long> _parsed_reads;
 		std::atomic<bool> _file_ended;
 
-		std::atomic<bool> _read_in_progress;
-		std::atomic<bool> _write_in_progress;
+		std::atomic<bool> _reading_in_progress;
 
-		Tools::BlockingConcurrentQueue<std::string> _records;
-		Tools::BlockingConcurrentQueue<std::string> _gzipped;
-
-		const std::shared_ptr<TextWriter> _writer;
+		const std::shared_ptr<ConcurrentGzWriter> _fastq_writer;
+		std::shared_ptr<ConcurrentGzWriter> _params_writer;
 		std::vector<std::shared_ptr<FastQReader>> _fastq_readers;
 		s_counter_t _num_reads_per_cb;
 
@@ -63,11 +63,9 @@ namespace TagsSearch
 	private:
 		static std::string get_file_uid(long random_seed = -1);
 
-		bool get_next_record(FastQReader::FastQRecord& record);
+		bool get_next_record(FastQReader::FastQRecord& record, Tools::ReadParameters &params);
 
 		void read_bunch(size_t number_of_iterations = 10000, size_t records_bunch_size = 5000);
-		void gzip_all();
-		void write_all();
 		void run_thread();
 
 	protected:
@@ -78,8 +76,8 @@ namespace TagsSearch
 
 	public:
 		TagsFinderBase(const std::vector<std::string> &fastq_filenames,
-		               const boost::property_tree::ptree &processing_config, const std::shared_ptr<TextWriter> &writer,
-		               bool save_stats);
+		               const boost::property_tree::ptree &processing_config, const std::shared_ptr<ConcurrentGzWriter> &writer,
+		               bool save_stats, bool save_read_params);
 
 		void run(int number_of_threads);
 		const s_counter_t& num_reads_per_cb() const;