+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.6.8] - 2017-09-22
+### Added
+* Filtration of reads by barcode quality ("*TagsSearch/Processing/min_barcode_quality*" and 
+"*Estimation/Other/min_barcode_quality*" fields in the config)
+* dropEst is now able to parse read type (e.g. exonic/intronic) from .bam file (see *config_desc.xml*)
+
 ## [0.6.7] - 2017-09-13
 ### Changed
 * Fixed bug, which led to erroneous parsing of incorrect read (e.g. reads without spacer for Indrop V1)